Add main kaggle script for santander data set to clean, modify, analyze, train, and predict customer satisfaction.
This script can be generalized to any dataset but was made for the Santander competition. It contains a kaggle_data class that that makes it easy to break a input data into training, test, and cv; add new columns; cleanup/remove columns; memmap data; output results in various formats; and run several algorithms against the data.

The script can choose which algorithms give best results and contains implementations of kmeans clustering, linear regression, logistic regression, RFC, neural nets, and boosting.

This class is meant as a quick wrapper to input a dataset and do initial exploration using standard python libraries.